Reuse code in non-tz functions
