+++ core/src/main/java/io/questdb/griffin/engine/functions/date/TodayWithTimezoneFunctionFactory.java
@@ -32,15 +32,11 @@ import io.questdb.griffin.FunctionFactory;
 import io.questdb.griffin.PlanSink;
 import io.questdb.griffin.SqlException;
 import io.questdb.griffin.SqlExecutionContext;
-import io.questdb.griffin.engine.functions.IntervalFunction;
-import io.questdb.griffin.engine.functions.UnaryFunction;
-import io.questdb.std.*;
-import io.questdb.std.datetime.TimeZoneRules;
-import io.questdb.std.datetime.microtime.TimestampFormatUtils;
+import io.questdb.std.IntList;
+import io.questdb.std.Interval;
+import io.questdb.std.ObjList;
 import io.questdb.std.datetime.microtime.Timestamps;
 import org.jetbrains.annotations.NotNull;
-
-import static io.questdb.std.datetime.TimeZoneRuleFactory.RESOLUTION_MICROS;
 
 public class TodayWithTimezoneFunctionFactory implements FunctionFactory {
     private static final String SIGNATURE = "today(S)";
